/cc @xiang90 

This adds build constraints in order to pass memory-map flags to bolt.Option.
If backend package passes syscall.MAP_POPULATE flag, the boltdb does read-ahead
which speeds up entire-database read, which then leads to faster storage
Restore. Benchmark result shows for 4GB, it opens and loads 6x faster in SSD,
and 12x faster in HDD. For 2GB, 1.6x faster with MAP_POPULATE in SSD.


<br>
<hr>
##### Background

storage package should do read ahead to make kernal preload some pages into
page cache and speed up the recover reads in etcd.

Related issue:
- https://github.com/coreos/etcd/issues/3786

<br>
##### Test script

```
	start := time.Now()
	kv := storage.New("test.db", &fg{})
	kv.Restore()
	fmt.Println("kv.Restore took:", time.Since(start))
```


<br>
##### Results: SSD (local machine)

```
Ubuntu 15.10 with SSD
model name  : Intel(R) Core(TM) i7-4910MQ CPU @ 2.90GHz
go version go1.5.1 linux/amd64
linux kernel version: 4.2.0-18-generic

open + load(restore) 4GB data of 1 million random bytes keys:
WITHOUT MAP_POPULATE: 2 min 5 sec (open 3.706302ms, restore 1m58.015088268s)
WITH    MAP_POPULATE: 21 sec (open 7.771817481s, restore 13.6191382s)

open + load(restore) 2GB data of 1 million simple integer keys:
WITHOUT MAP_POPULATE: 18 sec (open 10.893178ms, restore 17.197815402s)
WITH    MAP_POPULATE: 11 sec (open 3.923564768s, restore 6.810750348s)
```


<br>
##### Results: HDD (GCE)

```
Debian GNU/Linux 7.9 (wheezy) with HDD
machine type: n1-highmem-2 (2 vCPUs, 13 GB memory) / Intel Haswell
model name      : Intel(R) Xeon(R) CPU @ 2.30GHz
go version go1.5.1 linux/amd64
linux kernel version: 3.16.0-0.bpo.4-amd64

open + load(restore) 4GB data of 1 million random bytes keys:
WITHOUT MAP_POPULATE: 9 min 10 sec (open 28.758918ms, restore 8m40.518096833s)
WITH    MAP_POPULATE: 46 sec (open 36.944924626s, restore 9.460929924s)
```