@gyuho 2015-11-14
And I plan to buy a Window machine (I am using ubuntu machine now). Once I get it, I can try to support Windows as well.