@xiang90 2015-11-14
@gyuho Thanks for the extensive analysis. The code looks good to me.