@gyuho 2015-11-14
I added comments as you suggested. Please take a look again.

Thanks @xiang90