@gyuho 2015-11-14
@xiang90 

> (example: source_windows_amd64.go) where GOOS and GOARCH represent any known operating system and architecture values respectively, then the file is considered to have an implicit build constraint requiring those terms (in addition to any explicit constraints in the file). https://golang.org/pkg/go/build/#ArchChar

Explains if OS is specified by filename suffix, it will be included and the OSes with build tags `build +` are also added when building. So `boltoption_unix.go` would be included when compiled in unix. I added `linux` and `solaris` because unix is for some reason not included here https://github.com/golang/go/blob/master/src/go/build/syslist.go#L7 which is used to check the OS is good or not. But looking at Go standard libraries, it should be ok. I added it just to be safe. (https://github.com/golang/go/blob/f78a4c84ac8ed44aaf331989aa32e40081fd8f13/src/net/cgo_unix.go, https://github.com/golang/go/tree/73f329f47236c6952235db395802810156a7bd65/src/syscall)