@xiang90 2015-11-16
LGTM. Will merge when tests pass.